Move block.js and blockheader.js into lib/block/
