+++ lib/block/block.js
@@ -2,14 +2,14 @@
 
 var _ = require('lodash');
 var BlockHeader = require('./blockheader');
-var BN = require('./crypto/bn');
-var BufferUtil = require('./util/buffer');
-var BufferReader = require('./encoding/bufferreader');
-var BufferWriter = require('./encoding/bufferwriter');
-var Hash = require('./crypto/hash');
-var JSUtil = require('./util/js');
-var Transaction = require('./transaction');
-var $ = require('./util/preconditions');
+var BN = require('../crypto/bn');
+var BufferUtil = require('../util/buffer');
+var BufferReader = require('../encoding/bufferreader');
+var BufferWriter = require('../encoding/bufferwriter');
+var Hash = require('../crypto/hash');
+var JSUtil = require('../util/js');
+var Transaction = require('../transaction');
+var $ = require('../util/preconditions');
 
 /**
  * Instantiate a Block from a Buffer, JSON object, or Object with
