+++ lib/block/blockheader.js
@@ -1,12 +1,12 @@
 'use strict';
 
 var _ = require('lodash');
-var BN = require('./crypto/bn');
-var BufferUtil = require('./util/buffer');
-var BufferReader = require('./encoding/bufferreader');
-var BufferWriter = require('./encoding/bufferwriter');
-var Hash = require('./crypto/hash');
-var JSUtil = require('./util/js');
+var BN = require('../crypto/bn');
+var BufferUtil = require('../util/buffer');
+var BufferReader = require('../encoding/bufferreader');
+var BufferWriter = require('../encoding/bufferwriter');
+var Hash = require('../crypto/hash');
+var JSUtil = require('../util/js');
 
 /**
  * Instantiate a BlockHeader from a Buffer, JSON object, or Object with
