+++ lib/block/merkleblock.js
@@ -1,6 +1,6 @@
 'use strict';
 var _ = require('lodash');
-var BlockHeader = require('../blockheader');
+var BlockHeader = require('./blockheader');
 var BufferUtil = require('../util/buffer');
 var BufferReader = require('../encoding/bufferreader');
 var BufferWriter = require('../encoding/bufferwriter');
