+++ lib/transaction/transaction.js
@@ -21,7 +21,7 @@ var MultiSigScriptHashInput = Input.MultiSigScriptHash;
 var Output = require('./output');
 var Script = require('../script');
 var PrivateKey = require('../privatekey');
-var Block = require('../block');
+var Block = require('../block/block');
 var BN = require('../crypto/bn');
 
 var CURRENT_VERSION = 1;
